First save
I’m at a pretty good spot right now in the exam; the classes I need so
far are mostly implemented, and it runs, although it doesn’t do much
yet. Working on figuring out how to log in and where to create the
player, so I can start making the game work.